+++ trunk/src/headers/gx_main_interface.h
@@ -157,7 +157,7 @@ namespace gx_gui
     virtual void addJToggleButton(const char* label, float* zone);
     virtual void addCheckButton(const char* label, float* zone);
     virtual void addVerticalSlider(const char* label, float* zone, float init, float min, float max, float step);
-    virtual void addHorizontalSlider(const char* label, int* zone, float init, float min, float max, float step);
+    virtual void addHorizontalSlider(const char* label, float* zone, float init, float min, float max, float step);
     virtual void addHorizontalWheel(const char* label, float* zone, float init, float min, float max, float step);
     virtual void addregler(const char* label, float* zone, float init, float min, float max, float step);
     virtual void addbigregler(const char* label, float* zone, float init, float min, float max, float step);
